Wrap async in a class, prevent early free.
We need to wait for uv_close to complete before freeing the handle. At
the same time, the callback can be called after uv_close, but before
uv_close finishes.

So we allocate uv_async_t on the heap, along with the function, and zero
the function when the user of async is already gone. This way, we can
check for zero in spurious callbacks, and keep the uv_async_t itself
alive until after uv_close finishes.